// src/main/webapp/src/api/models/Committee.ts
/* tslint:disable */
/* eslint-disable */
/**
 * dicoop API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.13.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TimeSlot } from './TimeSlot';
import {
    TimeSlotFromJSON,
    TimeSlotFromJSONTyped,
    TimeSlotToJSON,
    TimeSlotToJSONTyped,
} from './TimeSlot';
import type { Person } from './Person';
import {
    PersonFromJSON,
    PersonFromJSONTyped,
    PersonToJSON,
    PersonToJSONTyped,
} from './Person';
import type { Settings } from './Settings';
import {
    SettingsFromJSON,
    SettingsFromJSONTyped,
    SettingsToJSON,
    SettingsToJSONTyped,
} from './Settings';

/**
 * 
 * @export
 * @interface Committee
 */
export interface Committee {
    /**
     * 
     * @type {string}
     * @memberof Committee
     */
    id?: string;
    /**
     * 
     * @type {Person}
     * @memberof Committee
     */
    evaluatedPerson?: Person;
    /**
     * 
     * @type {Date}
     * @memberof Committee
     */
    createdDate?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof Committee
     */
    useAvailability?: boolean;
    /**
     * 
     * @type {Settings}
     * @memberof Committee
     */
    settings?: Settings;
    /**
     * 
     * @type {TimeSlot}
     * @memberof Committee
     */
    timeSlot?: TimeSlot;
}

/**
 * Check if a given object implements the Committee interface.
 */
export function instanceOfCommittee(value: object): value is Committee {
    return true;
}

export function CommitteeFromJSON(json: any): Committee {
    return CommitteeFromJSONTyped(json, false);
}

export function CommitteeFromJSONTyped(json: any, ignoreDiscriminator: boolean): Committee {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'evaluatedPerson': json['evaluatedPerson'] == null ? undefined : PersonFromJSON(json['evaluatedPerson']),
        'createdDate': json['createdDate'] == null ? undefined : (new Date(json['createdDate'])),
        'useAvailability': json['useAvailability'] == null ? undefined : json['useAvailability'],
        'settings': json['settings'] == null ? undefined : SettingsFromJSON(json['settings']),
        'timeSlot': json['timeSlot'] == null ? undefined : TimeSlotFromJSON(json['timeSlot']),
    };
}

export function CommitteeToJSON(json: any): Committee {
    return CommitteeToJSONTyped(json, false);
}

export function CommitteeToJSONTyped(value?: Committee | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'evaluatedPerson': PersonToJSON(value['evaluatedPerson']),
        'createdDate': value['createdDate'] == null ? value['createdDate'] : value['createdDate'].toISOString(),
        'useAvailability': value['useAvailability'],
        'settings': SettingsToJSON(value['settings']),
        'timeSlot': TimeSlotToJSON(value['timeSlot']),
    };
}