// src/main/webapp/src/api/models/PersonType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * dicoop API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.13.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PersonType
 */
export interface PersonType {
    /**
     * 
     * @type {string}
     * @memberof PersonType
     */
    name?: string;
}

/**
 * Check if a given object implements the PersonType interface.
 */
export function instanceOfPersonType(value: object): value is PersonType {
    return true;
}

export function PersonTypeFromJSON(json: any): PersonType {
    return PersonTypeFromJSONTyped(json, false);
}

export function PersonTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): PersonType {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function PersonTypeToJSON(json: any): PersonType {
    return PersonTypeToJSONTyped(json, false);
}

export function PersonTypeToJSONTyped(value?: PersonType | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
    };
}