// src/main/webapp/src/api/models/Range.ts
/* tslint:disable */
/* eslint-disable */
/**
 * dicoop API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.13.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Range
 */
export interface Range {
    /**
     * 
     * @type {Array<number>}
     * @memberof Range
     */
    value?: Array<number>;
    /**
     * 
     * @type {number}
     * @memberof Range
     */
    min?: number;
    /**
     * 
     * @type {number}
     * @memberof Range
     */
    max?: number;
}

/**
 * Check if a given object implements the Range interface.
 */
export function instanceOfRange(value: object): value is Range {
    return true;
}

export function RangeFromJSON(json: any): Range {
    return RangeFromJSONTyped(json, false);
}

export function RangeFromJSONTyped(json: any, ignoreDiscriminator: boolean): Range {
    if (json == null) {
        return json;
    }
    return {
        
        'value': json['value'] == null ? undefined : json['value'],
        'min': json['min'] == null ? undefined : json['min'],
        'max': json['max'] == null ? undefined : json['max'],
    };
}

export function RangeToJSON(json: any): Range {
    return RangeToJSONTyped(json, false);
}

export function RangeToJSONTyped(value?: Range | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'value': value['value'],
        'min': value['min'],
        'max': value['max'],
    };
}